+++ maiko/src/internal/broker.rs
@@ -4,24 +4,26 @@ use tokio::{select, sync::mpsc::Receiver};
 use tokio_util::sync::CancellationToken;
 
 use super::Subscriber;
-use crate::{Envelope, Error, Event, Result, Topic};
+use crate::{Config, Envelope, Error, Event, Result, Topic};
 
-#[derive(Debug)]
 pub struct Broker<E: Event, T: Topic<E>> {
     receiver: Receiver<Arc<Envelope<E>>>,
     subscribers: Vec<Subscriber<E, T>>,
     cancel_token: Arc<CancellationToken>,
+    config: Arc<Config>,
 }
 
 impl<E: Event, T: Topic<E>> Broker<E, T> {
     pub fn new(
         receiver: Receiver<Arc<Envelope<E>>>,
         cancel_token: Arc<CancellationToken>,
+        config: Arc<Config>,
     ) -> Broker<E, T> {
         Broker {
             receiver,
             subscribers: Vec::new(),
             cancel_token,
+            config,
         }
     }
 
@@ -37,24 +39,60 @@ impl<E: Event, T: Topic<E>> Broker<E, T> {
         let topic = Topic::from_event(&e.event);
         self.subscribers
             .iter()
-            .filter(|s| s.topics.contains(&topic) && s.name != e.meta.actor_name().into())
+            .filter(|s| s.topics.contains(&topic))
+            .filter(|s| !s.sender.is_closed())
+            .filter(|s| !Arc::ptr_eq(&s.name, &e.meta.actor_name))
             .try_for_each(|subscriber| subscriber.sender.try_send(e.clone()))?;
         Ok(())
     }
 
     pub async fn run(&mut self) -> Result<()> {
+        let mut cleanup_interval = tokio::time::interval(self.config.maintenance_interval);
         loop {
             select! {
-                _ = self.cancel_token.cancelled() => {
-                    break;
-                }
+                _ = self.cancel_token.cancelled() => break,
                 Some(e) = self.receiver.recv() => {
                     self.send_event(&e)?;
                 },
-                else => break
+                _ = cleanup_interval.tick() => {
+                    self.cleanup();
+                }
             }
         }
+        self.shutdown().await;
         Ok(())
+    }
+
+    fn cleanup(&mut self) {
+        self.subscribers.retain(|s| !s.sender.is_closed());
+    }
+
+    async fn shutdown(&mut self) {
+        use tokio::time::*;
+
+        // Send messages that were in the queue at the time of shutdown
+        for _ in 0..self.receiver.len() {
+            if let Ok(e) = self.receiver.try_recv() {
+                let _ = self.send_event(&e); // Best effort
+            } else {
+                break; // Queue drained faster than expected
+            }
+        }
+
+        tokio::task::yield_now().await;
+
+        // Wait the inner channels to be consumed by the actors
+        let start = Instant::now();
+        let timeout = Duration::from_millis(10);
+        while !self.is_empty() && start.elapsed() < timeout {
+            sleep(Duration::from_micros(100)).await;
+        }
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.subscribers
+            .iter()
+            .all(|s| s.sender.is_closed() || s.sender.capacity() == s.sender.max_capacity())
     }
 }
 
@@ -89,8 +127,9 @@ mod tests {
     #[tokio::test]
     async fn test_add_subscriber() {
         let (tx, rx) = mpsc::channel(10);
+        let config = Arc::new(crate::Config::default());
         let cancel_token = Arc::new(CancellationToken::new());
-        let mut broker = Broker::<TestEvent, TestTopic>::new(rx, cancel_token);
+        let mut broker = Broker::<TestEvent, TestTopic>::new(rx, cancel_token, config);
         let subscriber =
             super::Subscriber::new(Arc::from("subscriber1"), &[TestTopic::A], tx.clone());
         assert!(broker.add_subscriber(subscriber).is_ok());